Address UI review notes
