+++ src/app/recipes/[id]/recipe-types.ts
@@ -1,0 +1,3 @@
+import type { RouterOutputs } from "~/trpc/react";
+
+export type Recipe = NonNullable<RouterOutputs["recipe"]["getRecipe"]>;
